test.py
from pylab import *
from rtlsdr import *

sdr = RtlSdr()

# configure device
sdr.sample_rate = 2.e6
sdr.center_freq = 145e6
sdr.gain = 4

samples = sdr.read_samples(256*1024)
print(samples[0], samples[200000], len(samples))
sdr.close()

# use matplotlib to estimate and plot the PSD
psd(samples, NFFT=1024, Fs=sdr.sample_rate/1e6, Fc=sdr.center_freq/1e6)
xlabel('Frequency (MHz)')
ylabel('Relative power (dB)')