requirements.txt
pyrtlsdr
SQLAlchemy
psycopg2-binary
matplotlib